#153 - symbols refactoring
