+++ src/Domain/HydraScript.Domain.IR/ISymbol.cs
@@ -2,7 +2,7 @@ namespace HydraScript.Domain.IR;
 
 public interface ISymbol
 {
-    public SymbolId Id { get; }
+    public ISymbolId<ISymbol> Id { get; }
     public string Name { get; }
     public Type Type { get; }
 }
