--- src/Domain/HydraScript.Domain.IR/Impl/SymbolIds/TypeSymbolId.cs
@@ -1,6 +1,0 @@
-namespace HydraScript.Domain.IR.Impl.SymbolIds;
-
-public class TypeSymbolId(string name) : SymbolId
-{
-    protected override string Value { get; } = "type " + name;
-}
