--- src/Domain/HydraScript.Domain.IR/Impl/SymbolIds/VariableSymbolId.cs
@@ -1,6 +1,0 @@
-namespace HydraScript.Domain.IR.Impl.SymbolIds;
-
-public class VariableSymbolId(string name) : SymbolId
-{
-    protected override string Value { get; } = "var " + name;
-}
