+++ src/Domain/HydraScript.Domain.IR/Impl/Symbols/Ids/TypeSymbolId.cs
@@ -1,0 +1,6 @@
+namespace HydraScript.Domain.IR.Impl.Symbols.Ids;
+
+public class TypeSymbolId(string name) : SymbolId<TypeSymbol>
+{
+    protected override string Value { get; } = "type " + name;
+}
