+++ src/Domain/HydraScript.Domain.IR/Impl/Symbols/Ids/VariableSymbolId.cs
@@ -1,0 +1,6 @@
+namespace HydraScript.Domain.IR.Impl.Symbols.Ids;
+
+public class VariableSymbolId(string name) : SymbolId<VariableSymbol>
+{
+    protected override string Value { get; } = "var " + name;
+}
