+++ src/Domain/HydraScript.Domain.IR/Impl/Symbols/TypeSymbol.cs
@@ -1,11 +1,11 @@
-using HydraScript.Domain.IR.Impl.SymbolIds;
+using HydraScript.Domain.IR.Impl.Symbols.Ids;
 
 namespace HydraScript.Domain.IR.Impl.Symbols;
 
 public class TypeSymbol(Type type, string? name = null) :
     Symbol(name ?? type.ToString(), type)
 {
-    public override SymbolId Id { get; } = new TypeSymbolId(name ?? type.ToString());
+    public override TypeSymbolId Id { get; } = new(name ?? type.ToString());
 
     public override bool Equals(object? obj) =>
         obj is TypeSymbol typeSymbol &&
